requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
docopt==0.6.2
# greenlet==0.4.17
greenlet
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
neovim==0.3.1
prompt-toolkit==3.0.8
pyflakes==2.2.0
Pygments==2.7.1
pylint==2.6.0
pynvim==0.4.2
pytz==2020.1
pyvim==3.0.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wcwidth==0.2.5
wrapt==1.12.1
